+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 1. 在线升级新增验证兼容最小版本号
 2. 【server】支持在线修改数据库账户密码
 3. 执行脚本文件由 `/bin/sh` 改为 `/bin/bash` 兼容 ubuntu 
+4. 【agent】项目 dsl 模式执行脚本变量支持直接引入 ${PROJECT_ID}、同时保留 #{PROJECT_ID} 引用
 
 ------